+++ src/lib.rs
@@ -4,7 +4,9 @@
 //!
 //! See [README](https://github.com/laysakura/succinct.rs/blob/master/README.md) for more about usage and features.
 
-pub use bit_vector::{BitVector, BitVectorBuilder, BitVectorString};
+pub use bit_string::BitString;
+pub use bit_vector::{BitVector, BitVectorBuilder};
 
+pub mod bit_string;
 pub mod bit_vector;
 mod internal_data_structure;