+++ tests/succinct_bit_vector_test.rs
@@ -1,4 +1,4 @@
-use succinct_rs::{BitVectorBuilder, BitVectorString};
+use succinct_rs::{BitString, BitVectorBuilder};
 
 #[test]
 fn build_from_length() {
@@ -19,15 +19,15 @@ fn build_from_length_and_set_bit() {
 }
 
 #[test]
-fn build_from_str() {
-    let bv = BitVectorBuilder::from_str(BitVectorString::new("01")).build();
+fn build_from_bit_string() {
+    let bv = BitVectorBuilder::from_bit_string(BitString::new("01")).build();
     assert_eq!(bv.access(0), false);
     assert_eq!(bv.access(1), true);
 }
 
 #[test]
-fn build_from_str_and_set_bit() {
-    let bv = BitVectorBuilder::from_str(BitVectorString::new("00"))
+fn build_from_bit_string_and_set_bit() {
+    let bv = BitVectorBuilder::from_bit_string(BitString::new("00"))
         .set_bit(0)
         .set_bit(1)
         .set_bit(0)
@@ -40,11 +40,11 @@ fn build_from_str_and_set_bit() {
 fn fuzzing_test() {
     let samples = 10000;
 
-    fn access_from_str(s: &str, i: u64) -> bool {
+    fn access_from_bit_string(s: &str, i: u64) -> bool {
         s.chars().collect::<Vec<char>>()[i as usize] == '1'
     }
 
-    fn rank_from_str(s: &str, i: u64) -> u64 {
+    fn rank_from_bit_string(s: &str, i: u64) -> u64 {
         let chs = s.chars().collect::<Vec<char>>();
         let mut rank: u64 = 0;
         for j in 0..=i as usize {
@@ -55,7 +55,7 @@ fn fuzzing_test() {
         rank
     }
 
-    fn select_from_str(s: &str, num: u64) -> Option<u64> {
+    fn select_from_bit_string(s: &str, num: u64) -> Option<u64> {
         if num == 0 {
             return Some(0);
         }
@@ -76,42 +76,42 @@ fn fuzzing_test() {
         let s = &format!("{:b}", rand::random::<u128>());
         eprintln!("build(): bit vec = \"{}\"", s);
 
-        let bvs = BitVectorString::new(s);
-        let bv = BitVectorBuilder::from_str(bvs).build();
+        let bs = BitString::new(s);
+        let bv = BitVectorBuilder::from_bit_string(bs).build();
 
         for i in 0..s.len() {
             eprintln!("access(): bit vec = \"{}\", i = {}, ", s, i);
             assert_eq!(
                 bv.access(i as u64),
-                access_from_str(s, i as u64),
-                "bit vec = \"{}\", i={}, BitVector::access()={}, access_from_str={}",
+                access_from_bit_string(s, i as u64),
+                "bit vec = \"{}\", i={}, BitVector::access()={}, access_from_bit_string={}",
                 s,
                 i,
                 bv.access(i as u64),
-                access_from_str(s, i as u64)
+                access_from_bit_string(s, i as u64)
             );
 
             eprintln!("rank(): bit vec = \"{}\", i = {}, ", s, i);
             assert_eq!(
                 bv.rank(i as u64),
-                rank_from_str(s, i as u64),
-                "bit vec = \"{}\", i={}, BitVector::rank()={}, rank_from_str={}",
+                rank_from_bit_string(s, i as u64),
+                "bit vec = \"{}\", i={}, BitVector::rank()={}, rank_from_bit_string={}",
                 s,
                 i,
                 bv.rank(i as u64),
-                rank_from_str(s, i as u64)
+                rank_from_bit_string(s, i as u64)
             );
 
             let num = i as u64;
             eprintln!("select(): bit vec = \"{}\", num = {}, ", s, num);
             assert_eq!(
                 bv.select(num),
-                select_from_str(s, num),
-                "bit vec = \"{}\", num={}, BitVector::select()={:?}, select_from_str={:?}",
+                select_from_bit_string(s, num),
+                "bit vec = \"{}\", num={}, BitVector::select()={:?}, select_from_bit_string={:?}",
                 s,
                 num,
                 bv.select(num),
-                select_from_str(s, num)
+                select_from_bit_string(s, num)
             );
         }
     }