+++ CHANGELOG.md
@@ -6,6 +6,16 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 
 ## [Unreleased]
 
+## [v0.2.0]
+
+### Added
+- `succinct_rs::BitString`
+- `succinct_rs::BitVectorBuilder::from_bit_string()`
+
+### Removed
+- `succinct_rs::BitVectorString`
+- `succinct_rs::BitVectorBuilder::from_str()`
+
 ## [v0.1.1]
 
 ### Fixed
@@ -14,8 +24,9 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 ## [v0.1.0]
 
 ### Added
-- `succinct::BitVector` and its builders: `succinct::BitVectorBuilder` and `succinct::BitVectorString`.
+- `succinct_rs::BitVector` and its builders: `succinct_rs::BitVectorBuilder` and `succinct_rs::BitVectorString`.
 
-[Unreleased]: https://github.com/laysakura/succinct.rs/compare/v0.1.1...HEAD
+[Unreleased]: https://github.com/laysakura/succinct.rs/compare/v0.2.0...HEAD
+[v0.2.0]: https://github.com/laysakura/succinct.rs/compare/v0.1.1...v0.2.0
 [v0.1.1]: https://github.com/laysakura/succinct.rs/compare/v0.1.0...v0.1.1
 [v0.1.0]: https://github.com/laysakura/succinct.rs/compare/3d425b4...v0.1.0